% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{add_expression_identifier}
\alias{add_expression_identifier}
\title{Title}
\usage{
add_expression_identifier(file_info, es_identifier)
}
\arguments{
\item{file_info}{}

\item{es_identifier}{}
}
\description{
Title
}

