% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_heatmap.R
\name{add_genomics_heatmap}
\alias{add_genomics_heatmap}
\title{add_genomics_heatmap}
\usage{
add_genomics_heatmap(p, mat, x = iheatmapr:::default_x(mat),
  col_order = c("groups", "none", "hclust", "kmeans"), col_groups = NULL,
  col_k = NULL, col_clust_dist = stats::dist, name = "Value",
  scale = c("rows", "cols", "none"), scale_method = c("standardize",
  "center", "normalize"), colorscale = continuous_colorscale(),
  show_colorbar = TRUE, col_groups_palette = iheatmapr:::DEFAULT_COLORS,
  col_groups_name = "Column<br>Groups", show_col_groups_colorbar = TRUE,
  x_labels = NULL, y_labels = NULL, x_title = NULL, y_title = NULL,
  buffer = 0.2, ...)
}
\arguments{
\item{mat}{matrix of values to be plotted as heatmap}

\item{x}{x xaxis labels, by default colnames of mat}

\item{col_order}{how to order columns? See Details}

\item{col_groups}{groups for columns, overriden if col_k is set and col_order is kmeans or hclust}

\item{col_k}{number of clusters for columns, needed if row_order is kmeans or optional if hclust}

\item{col_clust_dist}{distance function to use for column clustering if hierarchical clustering}

\item{name}{Name for colorbar}

\item{scale}{scale matrix by rows, cols or none}

\item{scale_method}{what method to use for scaling, either standardize or center}

\item{x_labels}{axis labels for x axis (default is x)}

\item{y_labels}{axis labels for y axis (default is NULL)}

\item{x_title}{x axis title}

\item{y_title}{y axis title}

\item{buffer}{amount of space to leave empty before this plot, relative to size 
of first heatmap}
}
\description{
add_genomics_heatmap
}
\author{
Alicia Schep
}

