% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{get_chip_file_info}
\alias{get_chip_file_info}
\title{get_chip_file_info}
\usage{
get_chip_file_info(samples_file, pairs_file)
}
\arguments{
\item{samples_file}{string}

\item{pairs_file}{string}
}
\value{
list
}
\description{
.. content for the "Description" section of the documentation ..
Make data.frame that specfies where necessary files are from NGS pipeline
.. content for "Details" section of the documentation ..
}
\author{
Justin Finkle
}

