% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_range.R
\name{get_symbol_range}
\alias{get_symbol_range}
\title{Get Symbol Range}
\usage{
get_symbol_range(db_object, symbol, keytype)
}
\arguments{
\item{exon.df}{data.frame: the exons included in the annotation}
}
\value{
range.gene GRange: s
}
\description{
Make a GRange that cover the exon list
}
\author{
Justin Finkle
}

