% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatory_potential.R
\name{get_tss_from_txdb}
\alias{get_tss_from_txdb}
\title{get_tss_from_txdb}
\usage{
get_tss_from_txdb(txdb, by_gene = TRUE, db = "org.Hs.eg")
}
\arguments{
\item{txdb}{txdb object}

\item{by_gene}{return a GRangesList with tss grouped by gene}

\item{db}{basename of meta-data package to use for \code{\link[annotate]{getSYMBOL}}}
}
\value{
GRanges or GRangesList
}
\description{
get_tss_from_txdb
}

