% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gviz_wrappers.R
\name{make_grtrack}
\alias{make_grtrack}
\title{Make GeneRegionTrack}
\usage{
make_grtrack(exon_data, genome, chr, kwargs = list())
}
\arguments{
\item{exon_data}{}

\item{genome}{}

\item{chr}{}

\item{kwargs}{}
}
\value{
grtrack
}
\description{
Make GeneRegionTrack
}

