% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{make_lines}
\alias{make_lines}
\title{Title}
\usage{
make_lines(symbol_range, scaled_range, hm_vals, base)
}
\arguments{
\item{symbol_range}{}

\item{scaled_range}{}

\item{hm_vals}{}

\item{base}{}
}
\description{
Title
}

