% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_heatmaps.R
\name{multi_coverage_heatmap}
\alias{multi_coverage_heatmap}
\title{multi_coverage_heatmap}
\usage{
multi_coverage_heatmap(mats, x = iheatmapr:::default_x(mats[[1]]),
  y = iheatmapr:::default_y(mats[[1]]), include = 1000,
  include_method = c("signal", "first", "random"), row_order = c("none",
  "hclust", "kmeans", "groups", "signal"), k = NULL, groups = NULL,
  clust_dist = stats::dist, cluster_by = c("first", "all"),
  signal = lapply(mats, rowMeans, na.rm = TRUE), plot_signal = TRUE,
  name = "Coverage", signal_name = "Avg. Coverage", summary = TRUE,
  scale_method = c("localRms", "localMean", "localNonZeroMean",
  "PercentileMax", "scalar", "none"), pct = 0.95, scale_factor = rep(1,
  length(mats)), show_xlabels = TRUE, start = x[1], end = default_end(x),
  xlab = "Position", font = list(size = 8), ...)
}
\arguments{
\item{mats}{list of coverage matrices}

\item{x}{x axis labels}

\item{y}{y axis labels}

\item{include}{how many rows to include in plot}

\item{include_method}{how to choose which rows to include in plot}

\item{row_order}{row order method}

\item{k}{k to use for kmeans clustering or cutting heirarchical clustering}

\item{groups}{pre-determined groups for rows}

\item{clust_dist}{distance function to use for clustering}

\item{cluster_by}{cluster using only first matrix or all matrices concatenated?}

\item{signal}{signal along row}

\item{plot_signal}{add an annotation heatmap showing the average signal? default is TRUE}

\item{name}{name of colorbar}

\item{summary}{make summary plot, boolean, default is TRUE}

\item{scale_method}{how to scale matrix before displaying in heatmap}

\item{pct}{percentile to use if scale_method is "PercentileMax"}

\item{scale_factor}{scale_factor to use if scale_method is "scalar"}

\item{show_xlabels}{show xlabels?  default is TRUE}

\item{start}{label for start of x range}

\item{end}{label for end of x range}

\item{xlab}{x axis label}

\item{font}{list of font attributes}
}
\value{
iheatmap object
}
\description{
makes multiple coverage heatmaps sharing the same scale
}
\author{
Alicia Schep
}

